Replace attribute NBT stuff with API
