+++ plugin/src/main/java/com/denizenscript/denizen/utilities/nbt/CustomNBT.java
@@ -1,9 +1,10 @@
 package com.denizenscript.denizen.utilities.nbt;
 
 import com.denizenscript.denizen.nms.NMSHandler;
-import com.denizenscript.denizen.nms.util.jnbt.*;
+import com.denizenscript.denizen.nms.util.jnbt.CompoundTag;
+import com.denizenscript.denizen.nms.util.jnbt.JNBTListTag;
+import com.denizenscript.denizen.nms.util.jnbt.StringTag;
 import com.denizenscript.denizen.objects.properties.entity.EntityDisabledSlots.Action;
-import com.denizenscript.denizencore.utilities.AsciiMatcher;
 import com.denizenscript.denizencore.utilities.CoreUtilities;
 import org.bukkit.Material;
 import org.bukkit.entity.Entity;
@@ -15,7 +16,6 @@ import java.util.*;
 public class CustomNBT {
 
     public static final String KEY_DENIZEN = "Denizen NBT";
-    public static final String KEY_ATTRIBUTES = "AttributeModifiers";
     public static final String KEY_CAN_PLACE_ON = "CanPlaceOn";
     public static final String KEY_CAN_DESTROY = "CanDestroy";
     public static final String KEY_DISABLED_SLOTS = "DisabledSlots";
@@ -36,149 +36,6 @@ public class CustomNBT {
      * Some static methods for dealing with Minecraft NBT data, which is used to store
      * custom NBT.
      */
-
-    public static class AttributeReturn {
-        public String attr;
-        public String slot;
-        public int op;
-        public double amt;
-        public long uuidMost;
-        public long uuidLeast;
-    }
-
-    public static List<AttributeReturn> getAttributes(ItemStack itemStack) {
-        if (itemStack == null || itemStack.getType() == Material.AIR) {
-            return null;
-        }
-        CompoundTag compoundTag = NMSHandler.itemHelper.getNbtData(itemStack);
-        List<CompoundTag> attribs = new ArrayList<>();
-        if (compoundTag.getValue().containsKey(KEY_ATTRIBUTES)) {
-            List<Tag> temp = (List<Tag>) compoundTag.getValue().get(KEY_ATTRIBUTES).getValue();
-            for (Tag tag : temp) {
-                attribs.add((CompoundTag) tag);
-            }
-        }
-        List<AttributeReturn> attrs = new ArrayList<>();
-        for (int i = 0; i < attribs.size(); i++) {
-            CompoundTag ct = attribs.get(i);
-            AttributeReturn atr = new AttributeReturn();
-            atr.attr = (String) ct.getValue().get("Name").getValue();
-            atr.slot = ct.getValue().get("Slot") == null ? "mainhand" : (String) ct.getValue().get("Slot").getValue();
-            atr.op = (Integer) ct.getValue().get("Operation").getValue();
-            Tag t = ct.getValue().get("Amount");
-            if (t instanceof IntTag) {
-                atr.amt = (Integer) t.getValue();
-            }
-            else if (t instanceof LongTag) {
-                atr.amt = (Long) t.getValue();
-            }
-            else if (t instanceof DoubleTag) {
-                atr.amt = (Double) t.getValue();
-            }
-            else {
-                /// ????
-                atr.amt = 0;
-            }
-            if (ct.getValue().containsKey("UUID")) {
-                UUID id = NMSHandler.itemHelper.convertNbtToUuid((IntArrayTag) ct.getValue().get("UUID"));
-                atr.uuidLeast = id.getLeastSignificantBits();
-                atr.uuidMost = id.getMostSignificantBits();
-            }
-            else if (ct.getValue().containsKey("UUIDMost")) {
-                t = ct.getValue().get("UUIDMost");
-                if (t instanceof LongTag) {
-                    atr.uuidMost = (Long) t.getValue();
-                }
-                else if (t instanceof IntTag) {
-                    atr.uuidMost = (Integer) t.getValue();
-                }
-                t = ct.getValue().get("UUIDLeast");
-                if (t instanceof LongTag) {
-                    atr.uuidLeast = (Long) t.getValue();
-                }
-                else if (t instanceof IntTag) {
-                    atr.uuidLeast = (Integer) t.getValue();
-                }
-            }
-            attrs.add(atr);
-        }
-        return attrs;
-    }
-
-    public static long uuidChoice(ItemStack its) {
-        String mat = CoreUtilities.toLowerCase(its.getType().name());
-        if (mat.contains("boots")) {
-            return 1000;
-        }
-        else if (mat.contains("legging")) {
-            return 100000;
-        }
-        else if (mat.contains("helmet")) {
-            return 10000000;
-        }
-        else if (mat.contains("chestp")) {
-            return 1000000000;
-        }
-        else {
-            return 1;
-        }
-    }
-
-    public static final AsciiMatcher uppercaseMatcher = new AsciiMatcher(AsciiMatcher.LETTERS_UPPER);
-
-    public static final HashMap<String, String> attributeNameUpdates = new HashMap<>();
-
-    static {
-        attributeNameUpdates.put("generic.maxHealth", "generic.max_health");
-        attributeNameUpdates.put("generic.followRange", "generic.follow_range");
-        attributeNameUpdates.put("generic.knockbackResistance", "generic.knockback_resistance");
-        attributeNameUpdates.put("generic.movementSpeed", "generic.movement_speed");
-        attributeNameUpdates.put("generic.flyingSpeed", "generic.flying_speed");
-        attributeNameUpdates.put("generic.attackDamage", "generic.attack_damage");
-        attributeNameUpdates.put("generic.attackKnockback", "generic.attack_knockback");
-        attributeNameUpdates.put("generic.attackSpeed", "generic.attack_speed");
-        attributeNameUpdates.put("generic.armorToughness", "generic.armor_toughness");
-    }
-
-    public static String fixAttributeName1_16(String input) {
-        if (!uppercaseMatcher.containsAnyMatch(input)) {
-            return input;
-        }
-        String replacement = attributeNameUpdates.get(input);
-        if (replacement != null) {
-            return replacement;
-        }
-        return CoreUtilities.toLowerCase(input);
-    }
-
-    public static ItemStack addAttribute(ItemStack itemStack, String attr, String slot, int op, double amt) {
-        if (itemStack == null || itemStack.getType() == Material.AIR) {
-            return null;
-        }
-        CompoundTag compoundTag = NMSHandler.itemHelper.getNbtData(itemStack);
-        List<CompoundTag> attribs = new ArrayList<>();
-        if (compoundTag.getValue().containsKey(KEY_ATTRIBUTES)) {
-            List<Tag> temp = (List<Tag>) compoundTag.getValue().get(KEY_ATTRIBUTES).getValue();
-            for (Tag tag : temp) {
-                attribs.add((CompoundTag) tag);
-            }
-        }
-        HashMap<String, Tag> tmap = new HashMap<>();
-        attr = fixAttributeName1_16(attr);
-        tmap.put("AttributeName", new StringTag(attr));
-        tmap.put("Name", new StringTag(attr));
-        tmap.put("Slot", new StringTag(slot));
-        tmap.put("Operation", new IntTag(op));
-        tmap.put("Amount", new DoubleTag(amt));
-        long uuidhelp = uuidChoice(itemStack);
-        UUID fullUuid = new UUID(uuidhelp + 88512 + attribs.size(), uuidhelp * 2 + 1250025L + attribs.size());
-        tmap.put("UUID", NMSHandler.itemHelper.convertUuidToNbt(fullUuid));
-        CompoundTag ct = NMSHandler.instance.createCompoundTag(tmap);
-        attribs.add(ct);
-        JNBTListTag lt = new JNBTListTag(CompoundTag.class, attribs);
-        compoundTag = compoundTag.createBuilder().put(KEY_ATTRIBUTES, lt).build();
-        return NMSHandler.itemHelper.setNbtData(itemStack, compoundTag);
-    }
 
     public static List<Material> getNBTMaterials(ItemStack itemStack, String key) {
         if (itemStack == null || itemStack.getType() == Material.AIR) {
